backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.1
aiosqlite==0.19.0
python-multipart==0.0.17
better-exceptions==0.3.2
pytest==8.3.4
httpx==0.27.2
pyjwt==2.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
cohere==5.11.0